pyproject.toml
[project]
name = "metal-price-gcp"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "google-cloud-storage (>=3.5.0,<4.0.0)",
    "google-cloud-bigquery (>=3.38.0,<4.0.0)",
    "google-cloud-core (>=2.5.0,<3.0.0)",
    "apache-airflow (>=3.1.2,<4.0.0)",
    "apache-airflow-providers-google (>=18.1.0,<19.0.0)",
    "apache-airflow-providers-http (>=5.4.0,<6.0.0)",
    "pyarrow (>=22.0.0,<23.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
